LLM_platform-main/app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
pymysql==1.1.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.2.0
httpx==0.27.0
openai==1.12.0